post-excerpt-ellipsis-suggestion
Hello, I have been using elementor for a couple years and would like to suggest adding an ellipsis option (the ...) at the end of post excerpts. It looks a bit sloppy having them just end abruptly. You can see an example if you scroll down a bit on the homepage of my site https://estnn.com/ .
If this is not the correct place to write a suggestion please let me know the correct spot, I was told by elementor support to post here. Keep up the good work developers :)